+++ packages/core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @web3uikit/core
 
+## 0.1.11
+
+### Patch Changes
+
+-   [#789](https://github.com/web3ui/web3uikit/pull/789) [`0a1aa35e9`](https://github.com/web3ui/web3uikit/commit/0a1aa35e9ca40077b6ff3beeaa85c466eb43f0b1) Thanks [@AbhinavMV](https://github.com/AbhinavMV)! - Add new small credit card and new credit card icon
+
+*   [#806](https://github.com/web3ui/web3uikit/pull/806) [`276c6e564`](https://github.com/web3ui/web3uikit/commit/276c6e56436060c0c52f9c3caed0ca6e67498f6e) Thanks [@AbhinavMV](https://github.com/AbhinavMV)! - [Button]: fix unique key warning in react
+
+*   Updated dependencies [[`0a1aa35e9`](https://github.com/web3ui/web3uikit/commit/0a1aa35e9ca40077b6ff3beeaa85c466eb43f0b1)]:
+    -   @web3uikit/icons@0.1.11
+
 ## 0.1.10
 
 ### Patch Changes